app folder restructure
