+++ src/Components/TestComponent.js
@@ -1,0 +1,7 @@
+
+import React from "react";
+
+const TestComponent =  () => (<h1>Test Component</h1>);
+
+
+export default TestComponent;
