+++ src/Services/TestService.js
@@ -1,0 +1,7 @@
+
+var TestService = async () => {
+    //make rest call to consume an API
+}
+
+
+export default TestService;
